djangocelery/requirements.backup.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
Babel==2.6.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.1.2
django-celery==3.2.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-guardian==1.4.9
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.0
flower==0.9.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
kombu==3.0.37
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.0
mongoengine==0.15.3
mysqlclient==1.3.13
pycparser==2.19
pymongo==3.7.1
pytz==2018.5
redis==2.10.6
requests==2.19.1
retrying==1.3.3
simplejson==3.16.0
six==1.11.0
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
